refactor(jsdoc): cookies
