+++ libs/@guardian/libs/src/cookies/ERR_INVALID_COOKIE.js
