+++ libs/@guardian/libs/src/cookies/cookies.test.ts
@@ -1,9 +1,9 @@
 import MockDate from 'mockdate';
-import { getCookie } from './getCookie';
-import * as getCookieValues from './getCookieValues';
-import { removeCookie } from './removeCookie';
-import { setCookie } from './setCookie';
-import { setSessionCookie } from './setSessionCookie';
+import { getCookie } from './getCookie.js';
+import * as getCookieValues from './getCookieValues.js';
+import { removeCookie } from './removeCookie.js';
+import { setCookie } from './setCookie.js';
+import { setSessionCookie } from './setSessionCookie.js';
 
 describe('cookies', () => {
 	let cookieValue = '';
@@ -11,9 +11,7 @@ describe('cookies', () => {
 	beforeAll(() => {
 		Object.defineProperty(document, 'cookie', {
 			get() {
-				return cookieValue
-					.replace('|', ';')
-					.replace(/^[;|]|[;|]$/g, '');
+				return cookieValue.replace('|', ';').replace(/^[;|]|[;|]$/g, '');
 			},
 
 			set(value: string) {
@@ -44,9 +42,7 @@ describe('cookies', () => {
 	it('gets a cookie', () => {
 		document.cookie =
 			'optimizelyEndUserId=oeu1398171767331r0.5280374749563634; __qca=P0-938012256-1398171768649;';
-		expect(getCookie({ name: '__qca' })).toEqual(
-			'P0-938012256-1398171768649',
-		);
+		expect(getCookie({ name: '__qca' })).toEqual('P0-938012256-1398171768649');
 	});
 
 	it('sets a cookie with an expiry date in six months that preserves UTC time', () => {
