+++ libs/@guardian/libs/src/cookies/getCookie.js
@@ -3,18 +3,13 @@ import { memoizedCookies } from './memoizedCookies';
 
 /**
  * Return a cookie. If it's been memoized it won't retrieve it again.
- * @param details Details about the cookie.
- * @param details.name - the cookie’s name.
- * @param details.shouldMemoize - set to true if you want to memoize it, default false.
+ * @param {object} details Details about the cookie.
+ * @param {string} details.name - the cookie’s name.
+ * @param {boolean} [details.shouldMemoize] - set to true if you want to memoize it, default false.
+ * @returns {string | null}
  */
 
-export const getCookie = ({
-	name,
-	shouldMemoize = false,
-}: {
-	name: string;
-	shouldMemoize?: boolean;
-}): string | null => {
+export const getCookie = ({ name, shouldMemoize = false }) => {
 	const memoizedCookie = memoizedCookies.get(name);
 	if (memoizedCookie) return memoizedCookie;
