+++ libs/@guardian/libs/src/cookies/getDomainAttribute.js
