+++ libs/@guardian/libs/src/cookies/getShortDomain.js
