+++ libs/@guardian/libs/src/cookies/memoizedCookies.js
@@ -1,0 +1,2 @@
+/** @type {Map<string, string>} */
+export const memoizedCookies = new Map();
