--- libs/@guardian/libs/src/cookies/memoizedCookies.ts
@@ -1,1 +1,0 @@
-export const memoizedCookies: Map<string, string> = new Map<string, string>();
