+++ libs/@guardian/libs/src/cookies/removeCookie.js
@@ -2,18 +2,12 @@ import { getShortDomain } from './getShortDomain';
 
 /**
  * Removes a cookie.
- * @param details Details about the cookie.
- * @param details.name - the cookie’s name.
- * @param details.currentDomainOnly - set to true if it's only for current domain
+ * @param {object} details Details about the cookie.
+ * @param {string} details.name - the cookie’s name.
+ * @param {boolean} [details.currentDomainOnly] - set to true if it's only for current domain
+ * @returns {void}
  */
-
-export const removeCookie = ({
-	name,
-	currentDomainOnly = false,
-}: {
-	name: string;
-	currentDomainOnly?: boolean;
-}): void => {
+export const removeCookie = ({ name, currentDomainOnly = false }) => {
 	const expires = 'expires=Thu, 01 Jan 1970 00:00:01 GMT;';
 	const path = 'path=/;';
