+++ libs/@guardian/libs/src/cookies/setSessionCookie.js
@@ -6,18 +6,12 @@ import { memoizedCookies } from './memoizedCookies';
 
 /**
  * Set a session cookie. If it's been memoized it will replace memoized value
- * @param details Details about the cookie.
- * @param details.name - the cookie’s name.
- * @param details.value - the cookie’s value.
+ * @param {object} details Details about the cookie.
+ * @param {string} details.name - the cookie’s name.
+ * @param {string} details.value - the cookie’s value.
+ * @returns {void}
  */
-
-export const setSessionCookie = ({
-	name,
-	value,
-}: {
-	name: string;
-	value: string;
-}): void => {
+export const setSessionCookie = ({ name, value }) => {
 	if (!isValidCookie(name, value)) {
 		throw new Error(`${ERR_INVALID_COOKIE} ${name}=${value}`);
 	}
